Move footnote form
